+++ src/handlers/handleFinish.ts
@@ -1,23 +1,34 @@
 import { AuditLevel } from 'src/types';
-import { printSecurityReport } from '../utils/print';
+import { printSecurityReport, printMaintainerReport } from '../utils/print';
 import { processAuditJson } from '../utils/vulnerability';
 
 /**
  * Process and analyze the NPM audit JSON
- * @param  {String} jsonBuffer    NPM audit stringified JSON payload
- * @param  {Number} auditLevel    The level of vulnerabilities we care about
- * @param  {Array} exceptionIds   List of vulnerability IDs to exclude
+ * @param  {String}   jsonBuffer          NPM audit stringified JSON payload
+ * @param  {Number}   auditLevel          The level of vulnerabilities we care about
+ * @param  {Array}    exceptionIds        List of vulnerability IDs to exclude
+ * @param  {Boolean}  shouldScanModules   Flag if we should scan the node_modules
  * @return {undefined}
  */
-export default function handleFinish(jsonBuffer: string, auditLevel: AuditLevel, exceptionIds: number[]): void {
-  const { unhandledIds, vulnerabilityIds, report, failed } = processAuditJson(jsonBuffer, auditLevel, exceptionIds);
+export default function handleFinish(
+  jsonBuffer: string,
+  auditLevel: AuditLevel,
+  exceptionIds: number[],
+  shouldScanModules?: boolean,
+): void {
+  const { unhandledIds, vulnerabilityIds, report, maintainerReport, failed } = processAuditJson(
+    jsonBuffer,
+    auditLevel,
+    exceptionIds,
+    shouldScanModules,
+  );
 
   // If unable to process the audit JSON
   if (failed) {
     console.error('Unable to process the JSON buffer string.');
     // Exit failed
     process.exit(1);
-    return;
+    return; // This seem unused but it is actually using in the test files to stop the process when we stubbing `process.exit()` above
   }
 
   // Print the security report
@@ -28,7 +39,7 @@ export default function handleFinish(jsonBuffer: string, auditLevel: AuditLevel,
   // Grab any un-filtered vulnerabilities at the appropriate level
   const unusedExceptionIds = exceptionIds.filter((id) => !vulnerabilityIds.includes(id));
 
-  // Display the unused exceptionId's
+  // Display the unused exception IDs
   if (unusedExceptionIds.length) {
     const messages = [
       `${
@@ -37,6 +48,12 @@ export default function handleFinish(jsonBuffer: string, auditLevel: AuditLevel,
       `${unusedExceptionIds.length > 1 ? 'They' : 'It'} can be removed from the .nsprc file or --exclude -x flags.`,
     ];
     console.warn(messages.join(' '));
+  }
+
+  // Display the auto excluded vulnerabilities
+  if (maintainerReport.length) {
+    printMaintainerReport(maintainerReport);
+    console.info('The auto scanning and exclusion is enabled by default, use `--scan-modules=false` to turn off this feature.');
   }
 
   // Display the found unhandled vulnerabilities