+++ src/handlers/handleInput.ts
@@ -9,7 +9,7 @@ import { getExceptionsIds } from '../utils/vulnerability';
  * @param  {Object} options     User's command options or flags
  * @param  {Function} fn        The function to handle the inputs
  */
-export default function handleInput(options: CommandOptions, fn: (T1: string, T2: AuditLevel, T3: number[]) => void): void {
+export default function handleInput(options: CommandOptions, fn: (T1: string, T2: AuditLevel, T3: number[], T4?: boolean) => void): void {
   // Generate NPM Audit command
   const auditCommand: string = [
     'npm audit',
@@ -29,5 +29,7 @@ export default function handleInput(options: CommandOptions, fn: (T1: string, T2
   const cmdExceptions: number[] = get(options, 'exclude', '').split(',').filter(isWholeNumber).map(Number);
   const exceptionIds: number[] = getExceptionsIds(nsprc, cmdExceptions);
 
-  fn(auditCommand, auditLevel, exceptionIds);
+  const shouldScanModules = options.scanModules !== 'false';
+
+  fn(auditCommand, auditLevel, exceptionIds, shouldScanModules);
 }